https://conduenttolling.visualstudio.com/Highways%20England/_workitems/edit/50200 Fixed: Application should provide correct chargers details for the vehicles mentioned     refer below correct details for charges .
https://conduenttolling.visualstudio.com/Highways%20England/_workitems/edit/50201
Fixed: Alignment and Format should be corrected and column of car , motorhomes and minibuses should displayed one by one without any continuation text.

https://conduenttolling.visualstudio.com/Highways%20England/_workitems/edit/49994
Fixed: Application is not displaying proper error message for Building number and street field which is accepting special characters which are not in requirement like {[$ so on.

https://conduenttolling.visualstudio.com/Highways%20England/_workitems/edit/50000
Fixed: Application is not displaying proper error message for  Address line 2 (optional) which is accepting special characters which are not in requirement like {[$ so on.